subir_preventiva/requirements.txt
pandas
impyla
python-dotenv 